Added C++ AI components
I implemented the AI in blueprints, and then converted each over to C++ one at a time, as I found this to be an easier approach than trying to implement everything from the ground up. This is due to the fact without everything working, it was hard to test if each individual class was doing what it was supposed to. I have also removed the Blueprint character that derived from the C++ character classes.